:bug: Enable xdebug in development
PHP xdebug was not yet enabled but also failed to work properly for me under
WSL2. host.docker.internal now correctly points to the host IP address, even
under linux since Docker 18.03 but with vscode remote-wsl, the xdebug server
is running from the linux kernel and the ports are not correctly shared between
the host machine and WSL2.

This commit introduces a workaround by exporting the WSL2 IP and passing it to
the app container via the extra_hosts configuration value inside the compose
file as per this SO question:

https://stackoverflow.com/questions/62104199/issues-when-debugging-php-in-vscode-using-docker-and-wsl2

Preferably, docker-for-win would setup forwarding properly as opposed to using
this workaround:

https://github.com/microsoft/WSL/issues/4150#issuecomment-504209723