fix: read only mounts result in incorrect file permissions
Inside the PHP container, the composer.json is mounted into a .config directory that is then owned by root instead of the docker user. This causes issues with the dotfiles/install.sh script. There are various solutions to this problem but I feel that it is ok for the development container to have read/write access to these files seeing as it's only being used in development and it should be possible to make the necessary modifications from inside the container when necessary.