refactor(php): simplify ondrej gpg key retrieval
Initially I made these changes due to a failing build but that turned
out to be caused by a failing DNS lookup. Somehow quad9 was resolving
just fine on the host machine but failed inside the container.

Keeping these since people have in the past reported issues with the
previous approach in laravel/sail which this file is loosely based upon.
And it's shorter.