style!: replace dot with dash notation
Kamal joins together the APP_NAME and service name with a dash and there
is as of yet no way to modify this. Technically we could use different
naming conventions for development and production but that will only
lead to confusion so I'm aligning things here.

docker-base will no longer be looking in the same place for your
container volumes so they will have to be renamed which, unfortunately,
is a bit involved. See
https://stackoverflow.com/questions/73233336/rename-docker-volume-on-docker-desktop
for more info.

BREAKING_CHANGE