feat(frankenphp): production container with Cloudflare DNS challenge
By default the container installs both the pgsql and mysql drivers. In
addition, some assumptions regarding where the files live (in a
subdirectory in the application project root) are done. The
.dockerignore file is included as an example and probably still needs to
be expanded upon.

I'll be using this as a base container running on a kubernetes cluster
or through docker-compose. Still needs testing in both contexts