refactor!: move and pull code from dedicated submodules
This is a major overhaul of the folder structure. From now on this
repository will serve as a reference and collection of all my Docker
files. Modules herein should instead be pulled into your project as a
submodule. The instructions for that can be found in the module's
repository README file.

BREAKING_CHANGE